requirements.txt
aiohttp-cors==0.7.0
numpy==1.24.4
pip-chill==1.0.3
pipreqs==0.5.0
pybind11==2.12.0
pycrypto==2.6.1
pycryptodome==3.20.0

setuptools~=75.1.0
six~=1.16.0
ipfshttpclient~=0.7.0
psutil~=5.9.0
pillow~=10.4.0
multiaddr~=0.0.9
urllib3~=2.2.3
requests~=2.32.3